scaffold out a uuid-macros crate that shared the parser
