+++ shared/README.md
@@ -1,0 +1,3 @@
+# Shared source
+
+This source is shared between `uuid` and `uuid-macros` to avoid circular dependencies or creating additional dependencies for `uuid`.
