+++ src/parser.rs
@@ -13,7 +13,10 @@
 //!
 //! [`Uuid`]: ../struct.Uuid.html
 
-use crate::{error::*, fmt, std::str, Uuid};
+use crate::{error::*, std::str, Uuid};
+
+#[path = "../shared/parser.rs"]
+mod imp;
 
 impl str::FromStr for Uuid {
     type Err = Error;
@@ -23,165 +26,14 @@ impl str::FromStr for Uuid {
     }
 }
 
-/// Check if the length matches any of the given criteria lengths.
-fn len_matches_any(len: usize, crits: &[usize]) -> bool {
-    for crit in crits {
-        if len == *crit {
-            return true;
-        }
-    }
-
-    false
-}
-
-// Accumulated length of each hyphenated group in hex digits.
-const ACC_GROUP_LENS: [usize; 5] = [8, 12, 16, 20, 32];
-
-// Length of each hyphenated group in hex digits.
-const GROUP_LENS: [usize; 5] = [8, 4, 4, 4, 12];
-
 impl Uuid {
     /// Parses a `Uuid` from a string of hexadecimal digits with optional
     /// hyphens.
     ///
     /// Any of the formats generated by this module (simple, hyphenated, urn)
     /// are supported by this parsing function.
-    pub fn parse_str(mut input: &str) -> Result<Uuid, Error> {
-        // Ensure length is valid for any of the supported formats
-        let len = input.len();
-
-        if len == fmt::Urn::LENGTH && input.starts_with("urn:uuid:") {
-            input = &input[9..];
-        } else if !len_matches_any(
-            len,
-            &[fmt::Hyphenated::LENGTH, fmt::Simple::LENGTH],
-        ) {
-            return Err(ErrorKind::InvalidLength {
-                expected: ExpectedLength::Any(&[
-                    fmt::Hyphenated::LENGTH,
-                    fmt::Simple::LENGTH,
-                ]),
-                found: len,
-            }
-            .into());
-        }
-
-        // `digit` counts only hexadecimal digits, `i_char` counts all chars.
-        let mut digit = 0;
-        let mut group = 0;
-        let mut acc = 0;
-        let mut buffer = [0u8; 16];
-
-        for (i_char, chr) in input.bytes().enumerate() {
-            if digit as usize >= fmt::Simple::LENGTH && group != 4 {
-                if group == 0 {
-                    return Err(ErrorKind::InvalidLength {
-                        expected: ExpectedLength::Any(&[
-                            fmt::Hyphenated::LENGTH,
-                            fmt::Simple::LENGTH,
-                        ]),
-                        found: len,
-                    }
-                    .into());
-                }
-
-                return Err(ErrorKind::InvalidGroupCount {
-                    expected: ExpectedLength::Any(&[1, 5]),
-                    found: group + 1,
-                }
-                .into());
-            }
-
-            if digit % 2 == 0 {
-                // First digit of the byte.
-                match chr {
-                    // Calculate upper half.
-                    b'0'..=b'9' => acc = chr - b'0',
-                    b'a'..=b'f' => acc = chr - b'a' + 10,
-                    b'A'..=b'F' => acc = chr - b'A' + 10,
-                    // Found a group delimiter
-                    b'-' => {
-                        if ACC_GROUP_LENS[group] as u8 != digit {
-                            // Calculate how many digits this group consists of
-                            // in the input.
-                            let found = if group > 0 {
-                                digit - ACC_GROUP_LENS[group - 1] as u8
-                            } else {
-                                digit
-                            };
-
-                            return Err(ErrorKind::InvalidGroupLength {
-                                expected: ExpectedLength::Exact(
-                                    GROUP_LENS[group],
-                                ),
-                                found: found as usize,
-                                group,
-                            }
-                            .into());
-                        }
-                        // Next group, decrement digit, it is incremented again
-                        // at the bottom.
-                        group += 1;
-                        digit -= 1;
-                    }
-                    _ => {
-                        return Err(ErrorKind::InvalidCharacter {
-                            expected: "0123456789abcdefABCDEF-",
-                            found: input[i_char..].chars().next().unwrap(),
-                            index: i_char,
-                            urn: UrnPrefix::Optional,
-                        }
-                        .into());
-                    }
-                }
-            } else {
-                // Second digit of the byte, shift the upper half.
-                acc *= 16;
-                match chr {
-                    b'0'..=b'9' => acc += chr - b'0',
-                    b'a'..=b'f' => acc += chr - b'a' + 10,
-                    b'A'..=b'F' => acc += chr - b'A' + 10,
-                    b'-' => {
-                        // The byte isn't complete yet.
-                        let found = if group > 0 {
-                            digit - ACC_GROUP_LENS[group - 1] as u8
-                        } else {
-                            digit
-                        };
-
-                        return Err(ErrorKind::InvalidGroupLength {
-                            expected: ExpectedLength::Exact(GROUP_LENS[group]),
-                            found: found as usize,
-                            group,
-                        }
-                        .into());
-                    }
-                    _ => {
-                        return Err(ErrorKind::InvalidCharacter {
-                            expected: "0123456789abcdefABCDEF-",
-                            found: input[i_char..].chars().next().unwrap(),
-                            index: i_char,
-                            urn: UrnPrefix::Optional,
-                        }
-                        .into());
-                    }
-                }
-                buffer[(digit / 2) as usize] = acc;
-            }
-            digit += 1;
-        }
-
-        // Now check the last group.
-        if ACC_GROUP_LENS[4] as u8 != digit {
-            return Err(ErrorKind::InvalidGroupLength {
-                expected: ExpectedLength::Exact(GROUP_LENS[4]),
-                found: (digit as usize - ACC_GROUP_LENS[3]),
-                group,
-            }
-            .into());
-        }
-
-        Ok(Uuid::from_bytes(buffer))
+    pub fn parse_str(input: &str) -> Result<Uuid, Error> {
+        Ok(Uuid::from_bytes(imp::parse_str(input)?))
     }
 }
